Added module to let admins disable xp on players
Admins can use the admin menu to toggle experience gaining for players
or use the sm_togglexp admin command.

This can be used to punish players, who are violating some rules and
farm exp with an unfair advantage.

You can list the status of experience for all players with the
sm_listdisabledexp command.