Try to fix crash when toggling turbo mode in menu
There seems to be a double free bug in SM's menu handling. When the
client, which selected a menu item is disconnected during that callback,
the menu is destroyed in the ondisconnect callback. Afterwards the menu
manager tries to destroy the menu normally in the menuend handler?

Program terminated with signal SIGSEGV, Segmentation fault.
#0  0xea0919e1 in non-virtual thunk to
MenuManager::OnHandleDestroy(unsigned int, void*) ()