Work around error spam on failed database connection
SourceMod 1.7 and previous don't handle calls to SetFailState correctly
and leave dependent plugins running while the parent plugin is stopped.

If the database failed to connect when the plugin was loaded, the core
plugin would enter an error state. All other plugins like upgrades still
tried to call core's natives like SMRPG_IsEnabled and errored, because
the plugin owning that native isn't running.

This patch works around it, by not entering an error state if the
connection fails, but just logging an error and trying again later. #95
#73 #26