Fix not enabling upgrades right away in BuySell callback
When loading upgrades from the database after the player was already put
in the server, the upgrade level was set before the |enabled| flag was
loaded from the database. The upgrade's buy callback was called while
the player had the upgrade "disabled", so he always had it on level 0.

This let to problems with the HP regeneration upgrade, where it won't
start regenerating after map changes etc.