Fix upgrade plugins being unable to get the current level in Buy callback
When upgrade levels of a client get loaded after he's ingame already,
the upgrade plugins were unable to get the current upgrade level in the
first Buy callback, due to the player being in the middle of the process
of loading all data from the database and SMRPG_GetClientUpgradeLevel
checking, if the client is fully loaded yet.

This could have some backdraws when upgrades try to get information of
other upgrades - see comment in commit.