Save stats and upgrade levels both at once on disconnect
When there is a lot of lag on the server and players reset their stats
during that time, it might happen, that the level is reset to 1 when
they rejoin, but the upgrade levels still stay the same.
Make sure both queries run to save the player's stats when he
disconnects.