Fix upgrade permission checks
Upgrades affecting other players checked for the victim's upgrade access
instead of the player who has the upgrade and uses it on the victim.
This lead to upgrades like firepistol not working when restricted to an
admin flag.

This patch adds a second parameter to distinguish between the target and
the issuer of the effect and does the access checks on the issuer.

https://forums.alliedmods.net/showthread.php?t=286268