Make upgrade BuySell and ActiveQuery callbacks optional
Don't require the SMRPG_UpgradeQueryCB and SMRPG_ActiveQueryCB when
registering an upgrade type. Upgrades requiring the callbacks might
register them using the new `SMRPG_SetUpgradeBuySellCallback` and
`SMRPG_SetUpgradeActiveQueryCallback` natives.

This change is backwards compatible. `SMRPG_RegisterUpgradeType` still
accepts the callbacks, but the new natives are the recommened way now.

This reduces code duplication in the upgrade plugins especially in the
active callback. That callback is now only called, if it is even
possible for the effect to be active on the client due to configs or
restrictions.