Upgrade Longjump: Fix autobunnyhopping and first jump boost #282
The player doesn't let go of `+jump` when `sv_autobunnyhopping` or a similar plugin is enabled. The jumping heuristic only checked for the time the player started to press that button and didn't boost the later auto-jumps.

The first jump in a series of jumps was supposed to get a higher boost than the jumps afterwards, but that boost was never applied.

The plugin handles `sv_autobunnyhopping` transparently now and offers a new convar `smrpg_ljump_autobunnyhopping` to enable that functionality if some other plugin provides the auto bunnyhopping feature.