Upgrade Long Jump: Add convar to select boost method
There are different ways to apply a speed boost on jump. THC:RPG has a buggy implementation which accidentally additionally doubles the player's speed instead of only increasing it by a certain percent.
It appears that players are used to this, so add a new convar `smrpg_ljump_boost_method` to let the server admin choose which speed boost method they like.

The current implementation in SM:RPG was technically incorrect too, but not in a noticable way in most cases. While keeping the default behavior as it was before (0), add another method (2) which obeys the velocity calculation in the Source engine completely. next to the buggy way THC:RPG did the boost (1).